Make ioctl() prototype match up with POSIX.

+++ src/include/sys/ioctl.h
@@ -31,7 +31,7 @@
 #define FIONREAD 1  // Returns the number of bytes available for reading.
 
 __BEGIN_DECLS
-int ioctl(int, unsigned long, ...);
+int ioctl(int, int, ...);
 __END_DECLS
 
 #endif
